+++ signature/src/dilithium_ed25519_signature.c
@@ -19,6 +19,7 @@
 
 #include "dilithium_type.h"
 #include "lc_ed25519.h"
+#include "lc_sha512.h"
 #include "ret_checkers.h"
 #include "visibility.h"
 
@@ -44,17 +45,104 @@ LC_INTERFACE_FUNCTION(int, lc_dilithium_ed25519_sign,
 		      size_t mlen, const struct lc_dilithium_ed25519_sk *sk,
 		      struct lc_rng_ctx *rng_ctx)
 {
+	uint8_t digest[LC_SHA512_SIZE_DIGEST];
 	int ret;
 
 	CKNULL(sig, -EINVAL);
 	CKNULL(sk, -EINVAL);
 
 	CKINT(lc_dilithium_sign(&sig->sig, m, mlen, &sk->sk, rng_ctx));
-	CKINT(lc_ed25519_sign(&sig->sig_ed25519, m, mlen, &sk->sk_ed25519,
-			      rng_ctx));
+
+	lc_hash(lc_sha512, m, mlen, digest);
+	CKINT(lc_ed25519ph_sign(&sig->sig_ed25519, digest, sizeof(digest),
+				&sk->sk_ed25519, rng_ctx));
+
+out:
+	lc_memset_secure(digest, 0, sizeof(digest));
+	return ret;
+}
+
+LC_INTERFACE_FUNCTION(int, lc_dilithium_ed25519_sign_init,
+		      struct lc_dilithium_ed25519_ctx *ctx,
+		      const struct lc_dilithium_ed25519_sk *sk)
+{
+	struct lc_hash_ctx *ed25519_hash_ctx;
+	int ret;
+
+	CKNULL(ctx, -EINVAL);
+	CKNULL(sk, -EINVAL);
+
+	ed25519_hash_ctx = &ctx->ed25519_hash_ctx;
+
+	/* Require the use of SHA-512 */
+	if (ed25519_hash_ctx->hash != lc_sha512)
+		return -EOPNOTSUPP;
+
+	CKINT(lc_dilithium_sign_init(&ctx->dilithium_hash_ctx, &sk->sk));
+
+	/* ED25519: Only perform hashing part */
+	lc_hash_init(ed25519_hash_ctx);
 
 out:
 	return ret;
+}
+
+LC_INTERFACE_FUNCTION(int, lc_dilithium_ed25519_sign_update,
+		      struct lc_dilithium_ed25519_ctx *ctx, const uint8_t *m,
+		      size_t mlen)
+{
+	struct lc_hash_ctx *ed25519_hash_ctx;
+	int ret;
+
+	CKNULL(ctx, -EINVAL);
+
+	ed25519_hash_ctx = &ctx->ed25519_hash_ctx;
+
+	CKINT(lc_dilithium_sign_update(&ctx->dilithium_hash_ctx, m, mlen));
+
+	/* ED25519: Only perform hashing part */
+	lc_hash_update(ed25519_hash_ctx, m, mlen);
+
+out:
+	return ret;
+}
+
+LC_INTERFACE_FUNCTION(int, lc_dilithium_ed25519_sign_final,
+		      struct lc_dilithium_ed25519_sig *sig,
+		      struct lc_dilithium_ed25519_ctx *ctx,
+		      const struct lc_dilithium_ed25519_sk *sk,
+		      struct lc_rng_ctx *rng_ctx)
+{
+	uint8_t digest[LC_SHA512_SIZE_DIGEST];
+	struct lc_hash_ctx *ed25519_hash_ctx;
+	int ret;
+
+	CKNULL(sig, -EINVAL);
+	CKNULL(sk, -EINVAL);
+	CKNULL(ctx, -EINVAL);
+
+	ed25519_hash_ctx = &ctx->ed25519_hash_ctx;
+
+	CKINT(lc_dilithium_sign_final(&sig->sig, &ctx->dilithium_hash_ctx,
+				      &sk->sk, rng_ctx));
+
+	lc_hash_final(ed25519_hash_ctx, digest);
+	CKINT(lc_ed25519ph_sign(&sig->sig_ed25519, digest, sizeof(digest),
+				&sk->sk_ed25519, rng_ctx));
+
+out:
+	lc_memset_secure(digest, 0, sizeof(digest));
+	return ret;
+}
+
+static inline int lc_dilithium_ed25519_verify_check(int retd, int rete)
+{
+	if (rete == -EBADMSG || retd == -EBADMSG)
+		return -EBADMSG;
+	else if (rete == -EINVAL || retd == -EINVAL)
+		return -EINVAL;
+	else
+		return rete | retd;
 }
 
 LC_INTERFACE_FUNCTION(int, lc_dilithium_ed25519_verify,
@@ -62,15 +150,123 @@ LC_INTERFACE_FUNCTION(int, lc_dilithium_ed25519_verify,
 		      const uint8_t *m, size_t mlen,
 		      const struct lc_dilithium_ed25519_pk *pk)
 {
-	int retd, rete;
+	uint8_t digest[LC_SHA512_SIZE_DIGEST];
+	int retd, rete, ret = 0;
+
+	CKNULL(sig, -EINVAL);
+	CKNULL(pk, -EINVAL);
 
 	retd = lc_dilithium_verify(&sig->sig, m, mlen, &pk->pk);
-	rete = lc_ed25519_verify(&sig->sig_ed25519, m, mlen, &pk->pk_ed25519);
 
-	if (rete == -EBADMSG || retd == -EBADMSG)
-		return -EBADMSG;
-	else if (rete == -EINVAL || retd == -EINVAL)
+	lc_hash(lc_sha512, m, mlen, digest);
+	rete = lc_ed25519ph_verify(&sig->sig_ed25519, digest, sizeof(digest),
+				   &pk->pk_ed25519);
+	lc_memset_secure(digest, 0, sizeof(digest));
+
+out:
+	return ret ? ret : lc_dilithium_ed25519_verify_check(retd, rete);
+}
+
+LC_INTERFACE_FUNCTION(int, lc_dilithium_ed25519_verify_init,
+		      struct lc_dilithium_ed25519_ctx *ctx,
+		      const struct lc_dilithium_ed25519_pk *pk)
+{
+	struct lc_hash_ctx *ed25519_hash_ctx;
+	int ret;
+
+	CKNULL(ctx, -EINVAL);
+	CKNULL(pk, -EINVAL);
+
+	ed25519_hash_ctx = &ctx->ed25519_hash_ctx;
+
+	/* Require the use of SHA-512 */
+	if (ed25519_hash_ctx->hash != lc_sha512)
+		return -EOPNOTSUPP;
+
+	CKINT(lc_dilithium_verify_init(&ctx->dilithium_hash_ctx, &pk->pk));
+
+	/* ED25519: Only perform hashing part */
+	lc_hash_init(ed25519_hash_ctx);
+
+out:
+	return ret;
+}
+
+LC_INTERFACE_FUNCTION(int, lc_dilithium_ed25519_verify_update,
+		      struct lc_dilithium_ed25519_ctx *ctx, const uint8_t *m,
+		      size_t mlen)
+{
+	struct lc_hash_ctx *ed25519_hash_ctx;
+	int ret;
+
+	CKNULL(ctx, -EINVAL);
+
+	ed25519_hash_ctx = &ctx->ed25519_hash_ctx;
+
+	CKINT(lc_dilithium_verify_update(&ctx->dilithium_hash_ctx, m, mlen));
+
+	/* ED25519: Only perform hashing part */
+	lc_hash_update(ed25519_hash_ctx, m, mlen);
+
+out:
+	return ret;
+}
+
+LC_INTERFACE_FUNCTION(int, lc_dilithium_ed25519_verify_final,
+		      const struct lc_dilithium_ed25519_sig *sig,
+		      struct lc_dilithium_ed25519_ctx *ctx,
+		      const struct lc_dilithium_ed25519_pk *pk)
+{
+	uint8_t digest[LC_SHA512_SIZE_DIGEST];
+	struct lc_hash_ctx *ed25519_hash_ctx;
+	int retd, rete, ret = 0;;
+
+	CKNULL(sig, -EINVAL);
+	CKNULL(pk, -EINVAL);
+	CKNULL(ctx, -EINVAL);
+
+	ed25519_hash_ctx = &ctx->ed25519_hash_ctx;
+
+	retd = lc_dilithium_verify_final(&sig->sig, &ctx->dilithium_hash_ctx,
+					 &pk->pk);
+
+	lc_hash_final(ed25519_hash_ctx, digest);
+	rete = lc_ed25519ph_verify(&sig->sig_ed25519, digest, sizeof(digest),
+				   &pk->pk_ed25519);
+	lc_memset_secure(digest, 0, sizeof(digest));
+
+out:
+	return ret ? ret : lc_dilithium_ed25519_verify_check(retd, rete);
+}
+
+LC_INTERFACE_FUNCTION(int, lc_dilithium_ed25519_ctx_alloc,
+		      struct lc_dilithium_ed25519_ctx **ctx)
+{
+	struct lc_dilithium_ed25519_ctx *out_ctx = NULL;
+	int ret;
+
+	if (!ctx)
 		return -EINVAL;
-	else
-		return rete | retd;
+
+	ret = lc_alloc_aligned((void **)&out_ctx, LC_HASH_COMMON_ALIGNMENT,
+			       LC_DILITHIUM_ED25519_CTX_SIZE);
+	if (ret)
+		return -ret;
+
+	LC_SHAKE_256_CTX((&(out_ctx)->dilithium_hash_ctx));
+	LC_SHA512_CTX((&(out_ctx)->ed25519_hash_ctx));
+
+	*ctx = out_ctx;
+
+	return 0;
+}
+
+LC_INTERFACE_FUNCTION(void, lc_dilithium_ed25519_ctx_zero_free,
+		      struct lc_dilithium_ed25519_ctx *ctx)
+{
+	if (!ctx)
+		return;
+
+	lc_dilithium_ed25519_ctx_zero(ctx);
+	lc_free(ctx);
 }